Ported C++ unit tests to TRIQS 3.0 and fixed problems revealed by them
* Link realevol_c to ezarpack in the PUBLIC mode. Also link it to
  ${arpack_ng_LIBRARIES}.

* There are missing includes and a `#pragma once` in
  `<triqs/utility/variant_extensions.hpp>`
  C.f. https://github.com/TRIQS/triqs/pull/799

* Fix a minor issue in `h5_read()` of `init_state`.

* Remove some stray references to `std14` namespace.

+++ c++/realevol/init_state.cpp
@@ -92,7 +92,7 @@ void h5_read(h5::group fg, std::string const &name, init_state & st) {
 
   double weight;
   h5_read(wst_gr, "weight", weight);
-  unsigned long sp_index;
+  int sp_index;
   h5_read(wst_gr, "sp_index", sp_index);
   st.weighted_states.emplace_back(state_on_subspace_t(st.sub_hilbert_spaces[sp_index]), weight);
   h5_read(wst_gr, "amplitudes", st.weighted_states.back().state.amplitudes());
